requirements.txt
django==3.2.3
djangorestframework
markdown
django-filter